requirements.txt
streamlit
tensorflow
scikit-learn
numpy
pandas
matplotlib
opencv-python
pillow
flask
werkzeug
scipy
seaborn
requests
scikit-image
joblib
pytest
pyyaml
argon2-cffi
argon2-cffi-bindings
arrow
asttokens
async-lru
attrs
Babel
beautifulsoup4
bleach
certifi
cffi
charset-normalizer
colorama
comm
contourpy
cycler
debugpy
decorator
defusedxml
executing
fastjsonschema
filelock
fonttools
fqdn
fsspec
h11
httpcore
httpx
idna
ipykernel
ipython
ipywidgets
isoduration
jedi
Jinja2
joblib
json5
jsonpointer
jsonschema
jsonschema-specifications
jupyter
jupyter-console
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab_pygments
jupyterlab_server
jupyterlab_widgets
kiwisolver
MarkupSafe
matplotlib-inline
mistune
mpmath
nbclient
nbconvert
nbformat
nest-asyncio
networkx
notebook
notebook_shim
platformdirs
prometheus-client
prompt-toolkit
psutil
pure-eval
pycparser
Pygments
pyparsing
python-dateutil
python-json-logger
pytz
pywin32
pywinpty
PyYAML
pyzmq
qtconsole
QtPy
referencing
rfc3339-validator
rfc3986-validator
rpds-py
ruamel.yaml
ruamel.yaml.clib
scikit-learn
torch
torchvision
tornado
tqdm
traitlets
types-python-dateutil
typing_extensions
tzdata
uri-template
urllib3
wcwidth
webcolors
webencodings
websocket-client
widgetsnbextension
